+++ src/commands/dice.ts
@@ -238,7 +238,9 @@ export default async function dice(
                     ?.content.replace(/[^\040-\176\200-\377]/gi, '')
                     .match(/^y(es)?/i)
             ) {
-                await awaitedMessage.first()?.delete();
+                if (awaitedMessage.first()?.deletable) {
+                    await awaitedMessage.first()?.delete();
+                }
                 answeredYes = true;
             }
         } catch {
